requirements.txt
absl-py==2.1.0
altair==5.1.2
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
bcrypt==4.1.2
blinker==1.6.3
boto3==1.34.58
botocore==1.34.58
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
comm==0.1.3
contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.7
decorator==4.4.2
distlib==0.3.8
dtw==1.4.0
easyocr==1.7.1
et-xmlfile==1.1.0
executing==1.2.0
fastdtw==0.3.4
filelock==3.12.4
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.28.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.60.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.31.6
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.24.0
ipympl==0.9.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab_widgets==3.0.10
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.42.0
lxml==4.9.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
music21==9.1.0
nest-asyncio==1.5.6
networkx==3.2
ninja==1.11.1.1
numba==0.59.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.20.0
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.1.1
pytest==8.0.2
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
s3transfer==0.10.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.11.3
seaborn==0.13.2
secret==0.8
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.22
stack-data==0.6.2
streamlit==1.28.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io==0.36.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
tf-keras==2.15.0
threadpoolctl==3.3.0
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
virtualenv==20.25.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
Werkzeug==3.0.0
widgetsnbextension==4.0.10
wrapt==1.14.1
zipp==3.17.0
